feat: use lua script

+++ corneast-core/src/main/java/com/alioth4j/corneast_core/pojo/QueryRespDTO.java
@@ -15,6 +15,6 @@ public class QueryRespDTO implements Serializable {
 
     private String key;
 
-    private Integer remainingTokenCount;
+    private Long remainingTokenCount;
 
 }
