+++ corneast-core/src/main/java/com/alioth4j/corneast_core/pojo/RegisterReqDTO.java
@@ -15,6 +15,6 @@ public class RegisterReqDTO implements Serializable {
 
     private String key;
 
-    private Integer tokenCount;
+    private Long tokenCount;
 
 }
